refactor: update rgbpp transfer params
