+++ packages/ckb/package.json
@@ -29,6 +29,7 @@
   },
   "devDependencies": {
     "@ckb-lumos/molecule": "0.22.0-next.5",
+    "@rgbpp-sdk/service": "workspace:^",
     "@types/lodash": "^4.14.202",
     "vitest": "^1.4.0"
   },
