+++ packages/ckb/src/index.ts
@@ -5,7 +5,6 @@ export * from './error';
 export * from './paymaster';
 export * from './types';
 export * from './rgbpp';
-export * from './spv';
 export * from './utils';
 export * from './queue';
 export * from './constants';
